+++ src/main/java/org/example/PaymentStrategy.java
@@ -1,5 +1,5 @@
 package org.example;
 
 public interface PaymentStrategy {
-    public String checkout(double amount);
+    public String checkout(int amount);
 }
